+++ Source/Objects/Hardware/Vme/GretinaTrigger/ORGretinaTriggerController.m
@@ -599,7 +599,7 @@
     }
 
     NSString* key = [NSString stringWithFormat: @"orca.ORGretinaTrigger%d.selectedtab",[model slot]];
-    int index = [tabView indexOfTabViewItem:tabViewItem];
+    NSInteger index = [tabView indexOfTabViewItem:tabViewItem];
     [[NSUserDefaults standardUserDefaults] setInteger:index forKey:key];
 	
 }
